+++ app/models/member.rb
@@ -1,0 +1,13 @@
+class Member < ActiveRecord::Base
+  # Concerns
+  include Deletable
+
+  # Model Validation
+  validates :first_name, :last_name, presence: true
+
+  # Model Methods
+
+  def name
+    "#{first_name} #{last_name}"
+  end
+end