+++ db/migrate/20151020194626_create_members.rb
@@ -1,0 +1,14 @@
+class CreateMembers < ActiveRecord::Migration
+  def change
+    create_table :members do |t|
+      t.string :first_name
+      t.string :last_name
+      t.string :email
+      t.string :phone
+      t.string :role
+      t.boolean :deleted, null: false, default: false
+
+      t.timestamps null: false
+    end
+  end
+end